Cleaning old website files
