+++ front/components/home/LandingLayout.tsx
@@ -15,12 +15,12 @@ import { useEffect, useState } from "react";
 import { useCookies } from "react-cookie";
 
 import RootLayout from "@app/components/app/RootLayout";
-import { A } from "@app/components/home/new/ContentComponents";
-import { FooterNavigation } from "@app/components/home/new/menu/FooterNavigation";
-import { MainNavigation } from "@app/components/home/new/menu/MainNavigation";
-import { MobileNavigation } from "@app/components/home/new/menu/MobileNavigation";
-import Particles, { shapeNamesArray } from "@app/components/home/new/Particles";
-import ScrollingHeader from "@app/components/home/new/ScrollingHeader";
+import { A } from "@app/components/home/ContentComponents";
+import { FooterNavigation } from "@app/components/home/menu/FooterNavigation";
+import { MainNavigation } from "@app/components/home/menu/MainNavigation";
+import { MobileNavigation } from "@app/components/home/menu/MobileNavigation";
+import Particles, { shapeNamesArray } from "@app/components/home/Particles";
+import ScrollingHeader from "@app/components/home/ScrollingHeader";
 import { ClientSideTracking } from "@app/lib/tracking/client";
 import { classNames } from "@app/lib/utils";
