+++ front/components/home/SolutionSection.tsx
@@ -2,7 +2,7 @@ import { Avatar } from "@dust-tt/sparkle";
 import type { ReactNode } from "react";
 import React from "react";
 
-import { H2, H5, P } from "@app/components/home/new/ContentComponents";
+import { H2, H5, P } from "@app/components/home/ContentComponents";
 import { classNames } from "@app/lib/utils";
 
 interface SolutionSectionProps {
