+++ front/components/home/content/Product/BlogSection.tsx
@@ -6,9 +6,9 @@ import {
   CarouselItem,
   CarouselNext,
   CarouselPrevious,
-} from "@app/components/home/new/Carousel";
-import { BlogBlock } from "@app/components/home/new/ContentBlocks";
-import { Grid, H2, P } from "@app/components/home/new/ContentComponents";
+} from "@app/components/home/Carousel";
+import { BlogBlock } from "@app/components/home/ContentBlocks";
+import { Grid, H2, P } from "@app/components/home/ContentComponents";
 import { classNames } from "@app/lib/utils";
 
 export function BlogSection() {
