+++ front/components/home/content/Product/FutureSection.tsx
@@ -2,14 +2,8 @@ import { Div3D, Hover3D } from "@dust-tt/sparkle";
 import Link from "next/link";
 import React from "react";
 
-import { ImgBlock } from "@app/components/home/new/ContentBlocks";
-import {
-  A,
-  Grid,
-  H2,
-  P,
-  Strong,
-} from "@app/components/home/new/ContentComponents";
+import { ImgBlock } from "@app/components/home/ContentBlocks";
+import { A, Grid, H2, P, Strong } from "@app/components/home/ContentComponents";
 import { classNames } from "@app/lib/utils";
 
 export function FutureSection() {
