+++ front/components/home/menu/MobileNavigation.tsx
@@ -8,13 +8,9 @@ import type { LinkProps } from "next/link";
 import Link from "next/link";
 import * as React from "react";
 
-import { menuConfig } from "@app/components/home/new/menu/config";
-import { ScrollArea } from "@app/components/home/new/ScrollArea";
-import {
-  Sheet,
-  SheetContent,
-  SheetTrigger,
-} from "@app/components/home/new/Sheet";
+import { menuConfig } from "@app/components/home/menu/config";
+import { ScrollArea } from "@app/components/home/ScrollArea";
+import { Sheet, SheetContent, SheetTrigger } from "@app/components/home/Sheet";
 import { classNames } from "@app/lib/utils";
 
 export function MobileNavigation() {
