+++ front/components/home/menu/config.ts
