+++ front/pages/index.tsx
@@ -5,8 +5,8 @@ import { BlogSection } from "@app/components/home/content/Product/BlogSection";
 import { FutureSection } from "@app/components/home/content/Product/FutureSection";
 import { IntroSection } from "@app/components/home/content/Product/IntroSection";
 import { TeamSection } from "@app/components/home/content/Product/TeamSection";
-import type { LandingLayoutProps } from "@app/components/home/new/LandingLayout";
-import LandingLayout from "@app/components/home/new/LandingLayout";
+import type { LandingLayoutProps } from "@app/components/home/LandingLayout";
+import LandingLayout from "@app/components/home/LandingLayout";
 import config from "@app/lib/api/config";
 import { getSession } from "@app/lib/auth";
 import { getUserFromSession } from "@app/lib/iam/session";
