+++ front/pages/security.tsx
@@ -5,14 +5,14 @@ import React from "react";
 import {
   HeaderContentBlock,
   ImgBlock,
-} from "@app/components/home/new/ContentBlocks";
-import { A, Grid, P, Strong } from "@app/components/home/new/ContentComponents";
-import type { LandingLayoutProps } from "@app/components/home/new/LandingLayout";
-import LandingLayout from "@app/components/home/new/LandingLayout";
+} from "@app/components/home/ContentBlocks";
+import { A, Grid, P, Strong } from "@app/components/home/ContentComponents";
+import type { LandingLayoutProps } from "@app/components/home/LandingLayout";
+import LandingLayout from "@app/components/home/LandingLayout";
 import {
   getParticleShapeIndexByName,
   shapeNames,
-} from "@app/components/home/new/Particles";
+} from "@app/components/home/Particles";
 import config from "@app/lib/api/config";
 import { getSession } from "@app/lib/auth";
 import { getUserFromSession } from "@app/lib/iam/session";
